search: query -> select, factor out backend parts of searchSpec
