+++ package.json
@@ -45,7 +45,8 @@
         "#replacer": "./src/replacer.js",
         "#reverse": "./src/reverse.js",
         "#search": "./src/search.js",
-        "#search-spec": "./src/common-util/search-spec.js",
+        "#search-shape": "./src/common-util/search-shape.js",
+        "#search-select": "./src/search-select.js",
         "#serialize": "./src/data/serialize.js",
         "#sort": "./src/common-util/sort.js",
         "#sugar": "./src/common-util/sugar.js",
